refactor: build scheduler, refactor guild.

+++ backend/repositories/guild/itemRecord.repository.js
@@ -1,8 +1,5 @@
-const Task = require('../../models/guild/task.model.js');
 const Item = require('../../models/guild/item.model.js');
 const ItemRecord = require('../../models/guild/itemRecord.model.js');
-const Adventurer = require('../../models/guild/adventurer.model.js');
-const User = require('../../models/user/user.model.js');
 const ApplicationError = require('../../utils/error/applicationError.js');
 
 class ItemRecordRepository {
@@ -28,6 +25,7 @@ class ItemRecordRepository {
 
   static async deleteAllByTaskAndUser(taskId, AdventurerId) {
     const items = await Item.getAll(taskId);
+    if (!items) return;
     await Promise.all(
       items.map(async ({ id: itemId }) => {
         await ItemRecord.deleteAllByItemAndUser(itemId, AdventurerId);
